requirements.txt
langchain==0.1.5
langchain-openai==0.0.3
langchain-community==0.0.13
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.29.0
faiss-cpu==1.7.4
altair<5
openai==1.7.1
streamlit-extras==0.3.6

